Update cartridge addresses to use uint16_t's

+++ include/cartridge.hpp
@@ -3,7 +3,7 @@
 class Cartridge {
     public:
         Cartridge();
-        std::vector<uint8_t> getCartridgeContents(int switchableBankStart, int switchableBankEnd);
+        std::vector<uint8_t> getCartridgeContents(uint16_t switchableBankStart, uint16_t switchableBankEnd);
     private:
         std::vector<uint8_t> fullCartridgeContents;
 };
